requirements.txt
# Flask Web Framework
Flask==3.0.0
Flask-CORS==4.0.0

# HTTP Requests
requests==2.31.0

# Speech Recognition
SpeechRecognition==3.10.0
pyaudio==0.2.14

# Audio Processing
numpy==1.26.2
scipy==1.12.0
sounddevice==0.4.6

# Computer Vision (for face detection)
opencv-python==4.9.0.80
Pillow==10.2.0
ultralytics==8.0.196
mediapipe==0.10.7

# File Processing
python-docx==1.1.0
PyPDF2==3.0.1

# Text Processing
nltk==3.8.1

# Environment Variables
python-dotenv==1.0.0

# Development Tools
pytest==7.4.3
black==23.12.1
flake8==7.0.0

# Optional: For advanced audio processing
librosa==0.10.1

# Optional: For better resume parsing
spacy==3.7.2

# Optional: For enhanced text processing
textblob==0.17.1
